Fix and test rewards calculation when altering the beneficiary
When a depositor alters their beneficiary, we need to snapshot both
the old beneficiary and the new beneficiary. Before we do either, we
need to snapshot the global rewards values. However, if we update the
global rewards values twice in the same call, it breaks the reward
calculation.

The Synthetix implementation assumed that only a single user's rewards
would be snapshot per call. To accomodate this requirement, we split the
global vs. user update into two methods, which is in some ways a preferable
implementatino regardless of this requirement. These are sensitive internal
methods that must be called in the proper conditions.