Rename vars and methods associated with staking mechanics for clarity
We move away from some of the variable and method names we inherited from
the synthetix staking contract in favor of more explicit and clear names
where possible. In particular, we introduce language around checkpointing
and accumulators to make it more obvious how the mechanism works. We also
add clarity around reward related methods to make it clear the rewards in
question are only unclaimed rewards.