Leave the scale factor into the unclaimed reward calculations
To avoid the accrual of precision loss as unclaimed rewards are calculated
and checkpointed, we leave the scale factor in the unclaimed reward
checkpoints. Our unclaimed reward method becomes a helper function only.
We also add a test to ensure this precision loss does not accrue to a
meaningful degree, even in extreme conditions.