In UniStaker tests, remember depositors and surrogates and prevent overlap
Due to a recent, undocumented change in the Foundry fuzzer implementation, we are now
seeing test failures where the address of a yet-to-be-deployed DelegationSurrogate is
selected as a depositor. This breaks our tests. This is occuring despite the include_storage
flag being set to false, which previously was sufficient for preventing this edge case.

To accomodate this, we implement a test helper that remembers each depositor and surrogate as
they're chosen and deployed respectively, and assures there is no overlap between the two
groups of addresses.